+++ packages/materials/src/index.ts
@@ -9,7 +9,7 @@ export type {
   SpriteTexture,
 } from "./sprite-material.js";
 export { SpriteMaterial } from "./sprite-material.js";
-export type { StandardMaterialOptions } from "./standard-material.js";
+export type { ColorRGB, StandardMaterialOptions } from "./standard-material.js";
 export { StandardMaterial } from "./standard-material.js";
 export type { MaterialTexture } from "./texture.js";
 export type { ColorRGBA, UnlitMaterialOptions } from "./unlit-material.js";